refactor(cli): separated test files according to their role
