+++ packages/cli-tool/tests/commands/configCommand.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import fs from 'fs-extra';
-import { logger } from '../src/utils/logger';
+import { logger } from '../../src/utils/logger';
 
 // Mock fs-extra
 vi.mock('fs-extra', () => ({
@@ -39,7 +39,7 @@ describe('loadConfig', () => {
   it('exits with code 1 when ignix.config.js does not exist', async () => {
     vi.mocked(fs.pathExists).mockResolvedValue(false as never);
 
-    const { loadConfig } = await import('../src/utils/config');
+    const { loadConfig } = await import('../../src/utils/config');
 
     try {
       await loadConfig();
@@ -64,7 +64,7 @@ describe('loadConfig', () => {
     vi.mocked(fs.pathExists).mockResolvedValue(true as never);
     vi.mocked(fs.readFile).mockResolvedValue(JSON.stringify(mockConfig) as never);
 
-    const { loadConfig } = await import('../src/utils/config');
+    const { loadConfig } = await import('../../src/utils/config');
     const result = await loadConfig();
 
     expect(result).toEqual(mockConfig);
@@ -76,7 +76,7 @@ describe('loadConfig', () => {
     vi.mocked(fs.writeFile).mockResolvedValue(undefined as never);
     vi.mocked(fs.remove).mockResolvedValue(undefined as never);
 
-    const { loadConfig } = await import('../src/utils/config');
+    const { loadConfig } = await import('../../src/utils/config');
 
     try {
       await loadConfig();
